GTK: prevent the user from creating invalid tikz
The free-form entry fields (graph element properties and node labels)
are a potential source of invalid tikz code.  Since we quote any
dodgy-looking text with { and }, we just need to make sure there are no
unmatched curly braces entered in those fields.

This will turn the entry widgets red when there are unmatched braces,
and refuse to make use of any such values.