Properly clear selection box
It turns out the selection box, rather than being properly set to the
zero rectangle, was being set to the value of an uninitialised NSRect
created on the stack.  This never caused any issues for me before,
because my system compiles things with _FORTIFY_SOURCE by default, which
zeros new variables.  It was only when I (indirectly) disabled this for
debugging in gdb that I started getting selection boxes appearing in odd
places.