When dragging, only snap leader node to grid
If you've carefully hand-positioned various nodes on a diagram relative
to each other, and then want to move the whole thing around, you
generally don't want to change their position relative to each other.

So we only snap the node being dragged to the grid.  Everything else in
the selection maintains its position relative to that node.