GTK: delay setting cursor on non-realised windows
Calling -[WidgetSurface setCursor:] when the underlying GtkWidget has
not been realised yet used to result in a warning that the GdkWindow did
not exist.  As a result, the cursor would not actually be set.

To deal with this, if the widget has not been realised, we delay setting
the cursor until it is.